[server] Do not use `arguments`
This was relying on a somewhat subtle feature that the arrow function
does not create the legacy `arguments` special variable.

Should also be easier for the compiler to infer since the rest argument
is simply an normal array and doesn't carry any information about the function call.
(`use strict` might have had a similar effect already.)